bug fix: Виправив помилку мапінгу об'єктів

+++ src/main/java/org/example/repository/BranchRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 
-public interface BranchRepository extends JpaRepository<Branch, Long>, JpaSpecificationExecutor {
+public interface BranchRepository extends JpaRepository<Branch, Integer>, JpaSpecificationExecutor {
 }
